// src/fees/entities/enums.ts
export enum PaymentPlan {
  MONTHLY = 'monthly',
  QUARTERLY = 'quarterly',
  YEARLY = 'yearly',
}

export enum FeeStatus {
  PENDING = 'pending',
  PARTIAL = 'partial',
  PAID = 'paid',
  OVERDUE = 'overdue',
}

export enum PaymentRecordStatus {
  PENDING = 'pending',
  PAID = 'paid',
  OVERDUE = 'overdue',
}